+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Function projects must now include a valid `project.toml` file to pass validation.
+  Functions generated using `sf generate function` already include this file.
+- The `context.org.data_api` client now uses the Salesforce REST API version specified by the
+  `com.salesforce.salesforce-api-version` key in `project.toml`, rather than using the Salesforce
+  Org's maximum supported REST API version. This version must be `'53.0'` or newer.
 - The error messages shown for invalid functions have been improved.
 
 ## [0.1.0] - 2022-12-14